C++11: enums -> enum class, xUSING_CONST

+++ Include/xLib/Debug/Profiler.cpp
@@ -28,7 +28,7 @@ xNAMESPACE_BEGIN2(xl, debug)
 //-------------------------------------------------------------------------------------------------
 Profiler::Profiler() :
     _isStarted(false),
-    _log      ( static_cast<std::size_t>(FileLog::LogSizes::lsDefaultMb) ),
+    _log      (FileLog::LogSizes::lsDefaultMb),
     _start    (0),
     _stop     (0),
     _duration (0)
